Raise the PHPStan analysis to level 8
Tighten the type surface across the clusterer, metadata, feed and service
layers so the static analysis passes at level 8 with zero errors and no
baseline. The bulk is type precision: list<> reindexing via array_values(),
int/float casts for entity identifiers and numeric coercions, defensive
is_scalar()/is_array() narrowing of untyped payload values before typed
parsers, and consolidating the day-summary shape behind the DaySummary alias
via inline @var narrowing where the pipeline mutates by reference.

Also removes two now-provably-dead code paths surfaced by the analysis: the
unused softDistanceEligible field/mutation on RunDetector's metadata map
(the separate $softDistanceEligible accumulator is the live tracker), and the
null-context storage in IndexLogEntry::normaliseContext() that contradicted
its own no-null return contract and the scalar-only toArray() context type.

+++ src/Service/Metadata/CompositeMetadataExtractor.php
@@ -86,7 +86,7 @@ final readonly class CompositeMetadataExtractor implements MetadataExtractorInte
             }
 
             $shouldCollectTelemetry = $this->configuration->shouldCollectTelemetry($extractor);
-            $startedAt              = $shouldCollectTelemetry ? hrtime(true) : null;
+            $startedAt              = $shouldCollectTelemetry ? (int) hrtime(true) : null;
 
             try {
                 if ($extractor->supports($filepath, $media) === false) {
